requirements.txt
bitsandbytes==0.49.2
gymnasium==1.2.3
huggingface_hub==0.36.2
minigrid==3.0.0
numpy==2.4.2
peft==0.10.0
Pillow==12.1.1
safetensors==0.7.0
torch==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.41.2
